added the maven to the project. fixed some bugs.
