+++ .idea/dataSources/f8f15ad8-bf4d-4a5c-8e58-e5d73ad7d41e/storage_v2/_src_/schema/main.uQUzAA.meta
@@ -1,0 +1,2 @@
+#n:main
+!<md> [0, 0, null, null, -2147483648, -2147483648]
