+++ src/main/java/View/UserCRUDView.java
@@ -1,0 +1,5 @@
+package View;
+
+public class UserCRUDView {
+    //TODO - make this class the view of the main window
+}
