+++ src/main/java/db/Managers/IDBManager.java
@@ -1,5 +1,7 @@
 package db.Managers;
 
+import db.DBResult;
+
 import java.sql.Connection;
 
 public interface IDBManager {
